+++ dataflow/common/logblocks.py
@@ -1,4 +1,4 @@
-def _log_start(logger, class_id):
+def log_start(logger, class_id):
     logger.info(f"")
     logger.info(f"")
     logger.info(f"")
@@ -6,7 +6,7 @@ def _log_start(logger, class_id):
     logger.info(f"{class_id} {_spacer} {class_id} start -->")
 
 
-def _log_end(logger, class_id):
+def log_end(logger, class_id):
     _spacer = "-" * 50
     logger.info(f"{class_id} {_spacer} <-- {class_id} end")
     logger.info(f"")