+++ dataflow/local_run/local_run.py
@@ -24,20 +24,21 @@ from dataflow.local_run.calls import run
 # # DATATYPE='processing'
 # FILEGROUPS = ['10_meteo']
 
-# CH-DAV
-SITE = 'ch-dav'
-DATATYPE = 'raw'
-# DATATYPE='processing'
-# FILEGROUPS = ['10_meteo', '11_meteo_hut', '12_meteo_forestfloor', '13_meteo_backup_eth',
-#               '13_meteo_nabel', '15_meteo_snowheight', '17_meteo_profile', '30_profile_ghg',
-#               '40_chambers_ghg']
-FILEGROUPS = ['30_profile_ghg']
-
-# # CH-FRU
-# SITE = 'ch-fru'
+# # CH-DAV
+# SITE = 'ch-dav'
 # DATATYPE = 'raw'
 # # DATATYPE='processing'
-# FILEGROUPS = ['10_meteo']
+# # FILEGROUPS = ['10_meteo', '11_meteo_hut', '12_meteo_forestfloor', '13_meteo_backup_eth',
+# #               '13_meteo_nabel', '15_meteo_snowheight', '17_meteo_profile', '30_profile_ghg',
+# #               '40_chambers_ghg']
+# FILEGROUPS = ['12_meteo_forestfloor']
+
+# CH-FRU
+SITE = 'ch-fru'
+DATATYPE = 'raw'
+# DATATYPE='processing'
+# FILEGROUPS = ['10_meteo', '13_meteo_pressure']
+FILEGROUPS = ['10_meteo']
 
 # # CH-LAE
 # SITE = 'ch-lae'
@@ -53,33 +54,33 @@ FILEGROUPS = ['30_profile_ghg']
 # FILEGROUPS = ['10_meteo']
 
 # # Processing Level-0
-# SITE = 'ch-aws'
+# # SITE = 'ch-aws'
 # # SITE = 'ch-cha'
 # # SITE = 'ch-dav'
 # # SITE = 'ch-das'
 # # SITE = 'ch-fru'
 # # SITE = 'ch-lae'
 # # SITE = 'ch-las'
-# # SITE = 'ch-oe2'
+# SITE = 'ch-oe2'
 # DATATYPE = 'processing'
 # FILEGROUPS = ['20_ec_fluxes']
 
 
-# Common
+# Common xxx
 SCRIPT = 'filescanner'
 ACCESS = 'server'
 DIRCONF = r'F:\Sync\luhk_work\20 - CODING\22 - POET\configs'
 # YEAR = list(range(2010, 2013))
-YEAR = 2023
-# MONTH = None
-MONTH = 4
+YEAR = 2005
+MONTH = None
+# MONTH = 9
 FILELIMIT = 0
 NEWESTFILES = 0
 # TESTUPLOAD = True
 TESTUPLOAD = False
-# N_ROWS = 10  # Only upload x number of rows of each file
+# N_ROWS = 100  # Only upload x number of rows of each file
 N_ROWS = None
-#
+
 kwargs = dict(script=SCRIPT, site=SITE, datatype=DATATYPE,
               access=ACCESS, dirconf=DIRCONF, year=YEAR,
               month=MONTH, filelimit=FILELIMIT, newestfiles=NEWESTFILES,